pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  {email = "BuriedInCode@tuta.io", name = "BuriedInCode"},
  {email = "bdpepple@gmail.com", name = "Brian Pepple"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Framework :: Pydantic :: 2",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Natural Language :: English",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: BSD",
  "Operating System :: POSIX :: Linux",
  "Operating System :: POSIX",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Internet"
]
dependencies = [
  "eval-type-backport >= 0.1.3; python_version < \"3.10\"",
  "pydantic >= 2.5.3",
  "ratelimit >= 2.2.1",
  "requests >= 2.26.0"
]
description = "Python wrapper for Metron API"
dynamic = ["version"]
keywords = ["comic", "comics", "metadata"]
license = {text = "GPL-3.0-or-later"}
maintainers = [
  {email = "BuriedInCode@tuta.io", name = "BuriedInCode"}
]
name = "mokkari"
readme = "README.md"
requires-python = ">= 3.8"

[project.optional-dependencies]
dev = [
  "pre-commit >= 3.5.0"
]
test = [
  "coverage[toml] >= 7.0",
  "pytest-cov >= 5.0.0",
  "requests-mock >= 1.9.3",
  "tox >= 4.11.3"
]

[project.urls]
"Bug Tracker" = "https://github.com/Metron-Project/mokkari/issues"
Homepage = "https://github.com/Metron-Project/mokkari"

[tool.coverage.report]
show_missing = true

[tool.coverage.run]
source = ["mokkari"]

[tool.hatch.version]
path = "mokkari/__init__.py"

[tool.pytest.ini_options]
addopts = ["--cov"]

[tool.ruff]
fix = true
line-length = 100
output-format = "grouped"
show-fixes = true
unsafe-fixes = true

[tool.ruff.format]
docstring-code-format = true
line-ending = "native"
skip-magic-trailing-comma = true

[tool.ruff.lint]
ignore = [
  "COM812",
  "COM812",
  "COM819",
  "D107",
  "D203",
  "D206",
  "D213",
  "D300",
  "E111",
  "E114",
  "E117",
  "E501",
  "EXE",
  "FBT",
  "ISC001",
  "ISC002",
  "PLR2004",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "S101",
  "TCH",
  "W191"
]
ignore-init-module-imports = true
select = ["ALL"]
task-tags = ["FIXME", "HACK", "TODO", "XXX", "http"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
mypy-init-return = true

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "pydantic.field_validator"]

[tool.ruff.lint.per-file-ignores]
"docs/*" = ["A001", "D104"]
"tests/*" = ["PLR2004", "SLF001", "T201", "T203"]
